requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.0.1
numpy==1.24.3
pandas>=0.19
Pillow==10.0.0
pickle-mixin==1.0.2
tensorflow==2.13.0
googletrans==4.0.0rc1
gTTS==2.3.2
ipython==8.4.0